chore: adjusted assessments visualization rules in strand reports
the previous rule for assessments visualization in strand reports was that assessment points without student entries should not be displayed. as we move towards the idea of showing strands planning to students and families prior to report cards release, we'll need to show the assessment point even if there's no student entry for it.

thinking about differentiation assessment points, we are now showing all the assessment points but we highlight the assessment points with entries (including moments assessments entries).

in this commit, we also implemented a rule to hide the final assessment when the parent report card is not visible to the user.

- adjusted `Assessments.list_strand_goals_student_entries/2` to list all the strand assessment points (even those without entries) and to ignore moments assessment points without entries
- added support to `prevent_preview` in `<.assessment_point_entry_display>` component