chore/wip: started strand details page adjustments (new UI, support to current cycle)
we started implementing the "new layout" using `<.neo_header>` component in strand live, and moving the placement of some things: added a rubric tab separating it from the assessments tab, moved the spiral visualization to moments tab, and placed cover in about tab.

functionality in rubrics, assessment, moments, and notes tab are broken due to changes in router.