refactor: removed `StrandRubric` in favor of new fields in `Rubric`
the `StrandRubric` is necessary only if we want to use the same rubric in more than one strand. this is something that is not happening right now, and this constraint is slowing down all development around this relationship between rubrics and strands. I'm embracing the idea that the rubric only exists in the context of a strand - no more "standalone" rubrics. maybe later we can add a "copy rubric" feature that would allow rubric reuse at the same time we avoid linking the same rubric to many different contexts. in the end, I think this is a better architecture/rule than it was

- skipping tests of pages we're not using (we need to cleanup later)