wip: implementing basic quizzes tab in strand moments
I'm experimenting with a new live view structure in this feature, trying to avoid using live components. instead of encapsulating the tab content in a live component, we are considering the tab content is a live view by itself, and we use a new component + hooks structure to handle the shared page UI and state.

- added moment quiz routes
- added support to `moment_id` opt in `Quizzes.list_quizzes/1`
- created `Quizzes.get_quiz/1`
- added support to `id` attr in `<.empty_state />`
- created `LantternWeb.QuizzesComponents` module, with `<.quizzes_list>` and `<.quiz_form_overlay>`
- created `QuizFormComponent`